Refactor order service to enhance order management functionality
- Updated OrderDataProviderAdapter to include new methods for creating orders, updating order status, and retrieving order reservation details.
- Introduced new response models: CreateOrderResponseModel, CreateOrderReservationResponseModel, OrderReservationDetailsResponseModel, and OrderDetailResponseModel.
- Refactored existing response models and presenters to align with new naming conventions and functionality.
- Added new input boundaries for creating orders, deleting reservations, and retrieving customer order details.
- Implemented interactor classes for handling business logic related to order creation, reservation management, and order status updates.
- Updated JpaOrder and JpaOrderItem entities to reflect changes in data structure and relationships.
- Enhanced repository interfaces to support new query methods for order and reservation management.
- Removed deprecated response model classes and adjusted imports accordingly.

+++ services/order-service/src/main/java/com/fcs/orders/entity/vo/OrderPayment.java
@@ -2,6 +2,6 @@ package com.fcs.orders.entity.vo;
 
 import java.time.LocalDate;
 
-public record OrderPayment(Integer cardNumber, Integer cvv, String cardHolderName, LocalDate exp, CardBanner banner) {
+public record OrderPayment(Long cardNumber, Integer cvv, String cardHolderName, LocalDate exp, CardBanner banner) {
 
 }
